+++ scripts/deployIdentityExample.ts
@@ -4,12 +4,12 @@ import path from 'path';
 import { OnchainIdentityDeployHelper } from '../test/helpers/OnchainIdentityDeployHelper';
 import { StateDeployHelper } from '../test/helpers/StateDeployHelper';
 import { deployPoseidons } from '../test/utils/deploy-poseidons.util';
-const pathOutputJson = path.join(__dirname, './deploy_output.json');
+
+// current state smart contract on opt-sepolia
+const STATE_ADDRESS = '0x9a1A258702050BcFB938Ad8Ec0996503473216d1';
+const PATH_OUTPUT_JSON = path.join(__dirname, './deploy_output.json');
 
 async function main() {
-  // current state smart contract on opt-sepolia
-  const stateAddress = '0x9a1A258702050BcFB938Ad8Ec0996503473216d1';
-
   const owner = (await ethers.getSigners())[0];
   const [poseidon2Elements, poseidon3Elements, poseidon4Elements] = await deployPoseidons(
     owner,
@@ -24,24 +24,22 @@ async function main() {
 
   const deployHelper = await OnchainIdentityDeployHelper.initialize([owner], true);
   const contracts = await deployHelper.deployIdentity(
-    stateAddress,
+    STATE_ADDRESS,
     smtLib,
     poseidon3Elements,
     poseidon4Elements
   );
 
-  const identity = contracts.identity;
-
   const outputJson = {
-    state: stateAddress,
+    state: STATE_ADDRESS,
     smtLib: await smtLib.getAddress(),
-    identity: await identity.getAddress(),
+    identity: await contracts.identity.getAddress(),
     poseidon2: await poseidon2Elements.getAddress(),
     poseidon3: await poseidon3Elements.getAddress(),
     poseidon4: await poseidon4Elements.getAddress(),
     network: process.env.HARDHAT_NETWORK
   };
-  fs.writeFileSync(pathOutputJson, JSON.stringify(outputJson, null, 1));
+  fs.writeFileSync(PATH_OUTPUT_JSON, JSON.stringify(outputJson, null, 1));
 }
 
 main()
